requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
tk==0.1.0
bcrypt==3.2.0
boto3==1.23.0
botocore==1.26.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django
django-allauth==0.50.0
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.3.4
flake8
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata
jmespath==1.0.0
mccabe 
oauthlib==3.2.0
Pillow==9.1.0
protobuf==3.20.1
psycopg2 
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle 
pycparser==2.21
pyflakes 
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sqlparse 
typing_extensions
uritemplate==4.1.1
urllib3==1.26.9
wincertstore==0.2
zipp 